Rework films distributors
